updating RLLib Wrapper and associated documentation
